+++ Runtime/Framework/ComposableBehaviour.cs
@@ -79,103 +79,6 @@ namespace DT.UniStart {
       return this.ioc.Add<T>(this.gameObject.AddComponent<T>());
     }
 
-    #region Helper Methods for IWatchable
-    /// <summary>
-    /// Watch a watchable for changes.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction Watch(IWatchable watchable, UnityAction action) {
-      watchable.AddListener(action);
-      this.onDestroy.AddListener(() => watchable.RemoveListener(action));
-      return action;
-    }
-    /// <summary>
-    /// Watch a watchable for changes.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction<T0> Watch<T0>(IWatchable<T0> watchable, UnityAction<T0> action) {
-      watchable.AddListener(action);
-      this.onDestroy.AddListener(() => watchable.RemoveListener(action));
-      return action;
-    }
-    /// <summary>
-    /// Watch a watchable for changes.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction<T0, T1> Watch<T0, T1>(IWatchable<T0, T1> watchable, UnityAction<T0, T1> action) {
-      watchable.AddListener(action);
-      this.onDestroy.AddListener(() => watchable.RemoveListener(action));
-      return action;
-    }
-    /// <summary>
-    /// Watch a watchable for changes.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction<T0, T1, T2> Watch<T0, T1, T2>(IWatchable<T0, T1, T2> watchable, UnityAction<T0, T1, T2> action) {
-      watchable.AddListener(action);
-      this.onDestroy.AddListener(() => watchable.RemoveListener(action));
-      return action;
-    }
-    /// <summary>
-    /// Watch a watchable for changes.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction<T0, T1, T2, T3> Watch<T0, T1, T2, T3>(IWatchable<T0, T1, T2, T3> watchable, UnityAction<T0, T1, T2, T3> action) {
-      watchable.AddListener(action);
-      this.onDestroy.AddListener(() => watchable.RemoveListener(action));
-      return action;
-    }
-    #endregion
-
-    #region Helper Methods for IEventBus
-    /// <summary>
-    /// Watch an event bus for events.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction Watch<K>(IEventBus<K> eventBus, K key, UnityAction action) {
-      eventBus.AddListener(key, action);
-      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
-      return action;
-    }
-    /// <summary>
-    /// Watch an event bus for events.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction<T0> Watch<K, T0>(IEventBus<K> eventBus, K key, UnityAction<T0> action) {
-      eventBus.AddListener(key, action);
-      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
-      return action;
-    }
-    /// <summary>
-    /// Watch an event bus for events.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction<T0, T1> Watch<K, T0, T1>(IEventBus<K> eventBus, K key, UnityAction<T0, T1> action) {
-      eventBus.AddListener(key, action);
-      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
-      return action;
-    }
-    /// <summary>
-    /// Watch an event bus for events.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction<T0, T1, T2> Watch<K, T0, T1, T2>(IEventBus<K> eventBus, K key, UnityAction<T0, T1, T2> action) {
-      eventBus.AddListener(key, action);
-      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
-      return action;
-    }
-    /// <summary>
-    /// Watch an event bus for events.
-    /// Remove the listener when the object is destroyed.
-    /// </summary>
-    public UnityAction<T0, T1, T2, T3> Watch<K, T0, T1, T2, T3>(IEventBus<K> eventBus, K key, UnityAction<T0, T1, T2, T3> action) {
-      eventBus.AddListener(key, action);
-      this.onDestroy.AddListener(() => eventBus.RemoveListener(key, action));
-      return action;
-    }
-    #endregion
-
-    #region Composable Events
     public AdvancedEvent<int> onAnimatorIK => this.GetOrAddComponent<ComposableAnimatorIK>().@event;
     public AdvancedEvent onAnimatorMove => this.GetOrAddComponent<ComposableAnimatorMove>().@event;
     public AdvancedEvent<bool> onApplicationFocus => this.GetOrAddComponent<ComposableApplicationFocus>().@event;
@@ -232,6 +135,5 @@ namespace DT.UniStart {
     public AdvancedEvent onUpdate => this.GetOrAddComponent<ComposableUpdate>().@event;
     public AdvancedEvent onValidate => this.GetOrAddComponent<ComposableValidate>().@event;
     public AdvancedEvent onWillRenderObject => this.GetOrAddComponent<ComposableWillRenderObject>().@event;
-    #endregion
   }
 }